+++ src/main/resources/db/changelog/scripts/V2__create_feature_flags_table.sql
@@ -1,9 +1,9 @@
 CREATE TABLE feature_flags (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
     description TEXT,
     enabled BOOLEAN DEFAULT false,
-    organization_id INTEGER REFERENCES organizations(id),
+    organization_id BIGINT REFERENCES organizations(id),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
